[pkg/workloadmeta/collectors] Move collectors to `internal` folder
