+++ .github/CODEOWNERS
@@ -211,7 +211,7 @@
 /pkg/appsec/                            @DataDog/agent-appsec
 /pkg/config/appsec.go                   @DataDog/agent-appsec
 /pkg/workloadmeta/                      @DataDog/container-integrations
-/pkg/workloadmeta/collectors/cloudfoundry @DataDog/integrations-tools-and-libraries
+/pkg/workloadmeta/collectors/internal/cloudfoundry @DataDog/integrations-tools-and-libraries
 
 /pkg-config/                            @DataDog/agent-platform
