+++ pkg/workloadmeta/collectors/internal/containerd/stub.go
