+++ pkg/workloadmeta/collectors/internal/docker/stub.go
