+++ pkg/workloadmeta/collectors/internal/ecs/stub.go
