+++ pkg/workloadmeta/collectors/internal/ecsfargate/stub.go
