+++ pkg/workloadmeta/collectors/internal/kubelet/stub.go
