+++ pkg/workloadmeta/collectors/internal/kubemetadata/stub.go
