+++ pkg/workloadmeta/collectors/internal/podman/stub.go
